+++ changelog.md
@@ -1,8 +1,14 @@
+## v 1.1.0 - 2022.10.XX
+* Alignment with upstream version [1.3.0 #d75bb2a] (https://github.com/adafruit/Adafruit_nRF52_Arduino/tree/d75bb2a8524e7f00f725eb2ab27a8996be65e7a3)
+* added reset via ota [#r50afx] (https://app.clickup.com/t/r50afx)
+* removed define baudrate `Arduino.h` [#37pztpm] (https://app.clickup.com/t/37pztpm)
+* Added FreeRTOS menu for consistency with other Arancino cores
+* Implemented thread-safe realloc and calloc
 
-## v 1.0.1 - 2022.07.11
-* removed task resume and suspend from `rtos.cpp` (#363bt7b)[https://app.clickup.com/t/363bt7b]
+## v 1.0.1 - 2022.07.XX
+* removed task resume and suspend from `rtos.cpp` [#363bt7b] (https://app.clickup.com/t/363bt7b)
 
-## v 1.0.0 - 2021.12.07
+## v 1.0.0 - 2021.09.XX
 
 * added arancino volante variant
 * added begin method in WString class